gridify update 2
Hi Ralf,

Thanks for the advice! I've updated squarify to combine with squarify2. 
For the seam artefact removal step, do you think it's a good idea to include a user input to switch this on an off? At the moment, it is switched on only if the unit cell length is specified.
This is useful for some applications (grains, plotting nice maps) but not others (GND density, KAM) and is computationally a bit expensive. If so, what's the best way to include a user switch into this function?

Thanks,
Vivian